test: Handle Singleton Display class
Fixes: #83538
